feat: export FacebookBaseConnector
