+++ packages/bottender/src/messenger/FacebookBaseConnector.ts
@@ -1,0 +1,192 @@
+import crypto from 'crypto';
+
+import invariant from 'invariant';
+import shortid from 'shortid';
+import { BatchConfig, MessengerBatchQueue } from 'messenger-batch';
+import { JsonObject } from 'type-fest';
+import { MessengerClient } from 'messaging-api-messenger';
+
+import { RequestContext } from '../types';
+
+type CommonConnectorOptions = {
+  appId: string;
+  appSecret: string;
+  verifyToken?: string;
+  batchConfig?: BatchConfig;
+  mapPageToAccessToken?: (pageId: string) => Promise<string>;
+};
+
+type ConnectorOptionsWithoutClient = {
+  ClientClass: typeof MessengerClient;
+  accessToken?: string;
+  origin?: string;
+  skipAppSecretProof?: boolean;
+} & CommonConnectorOptions;
+
+type ConnectorOptionsWithClient<C extends MessengerClient> = {
+  client: C;
+} & CommonConnectorOptions;
+
+export type FacebookBaseConnectorOptions<C extends MessengerClient> =
+  | ConnectorOptionsWithoutClient
+  | ConnectorOptionsWithClient<C>;
+
+export default class FacebookBaseConnector<
+  RequestBody extends JsonObject,
+  Client extends MessengerClient
+> {
+  _client: Client;
+
+  _appId: string;
+
+  _appSecret: string;
+
+  _mapPageToAccessToken: ((pageId: string) => Promise<string>) | null = null;
+
+  _verifyToken: string | null = null;
+
+  _batchConfig: BatchConfig | null = null;
+
+  _batchQueue: MessengerBatchQueue | null = null;
+
+  constructor(options: FacebookBaseConnectorOptions<Client>) {
+    const {
+      appId,
+      appSecret,
+      mapPageToAccessToken,
+      verifyToken,
+      batchConfig,
+    } = options;
+
+    if ('client' in options) {
+      this._client = options.client;
+    } else {
+      const { ClientClass, accessToken, origin, skipAppSecretProof } = options;
+
+      invariant(
+        accessToken || mapPageToAccessToken,
+        'Facebook access token is required. Please make sure you have filled it correctly in `bottender.config.js` or `.env` file.'
+      );
+      invariant(
+        appSecret,
+        'Facebook app secret is required. Please make sure you have filled it correctly in `bottender.config.js` or `.env` file.'
+      );
+
+      this._client = ClientClass.connect({
+        accessToken: accessToken || '',
+        appSecret,
+        origin,
+        skipAppSecretProof,
+      }) as Client;
+    }
+
+    this._appId = appId;
+    this._appSecret = appSecret;
+
+    this._mapPageToAccessToken = mapPageToAccessToken || null;
+    this._verifyToken = verifyToken || shortid.generate();
+
+    this._batchConfig = batchConfig || null;
+    if (this._batchConfig) {
+      this._batchQueue = new MessengerBatchQueue(
+        this._client,
+        this._batchConfig
+      );
+    }
+  }
+
+  get client(): Client {
+    return this._client;
+  }
+
+  get verifyToken(): string | null {
+    return this._verifyToken;
+  }
+
+  // https://developers.facebook.com/docs/messenger-platform/webhook#security
+  verifySignature(rawBody: string, signature: string): boolean {
+    if (typeof signature !== 'string') return false;
+
+    const sha1 = signature.split('sha1=')[1];
+
+    if (!sha1) return false;
+
+    const bufferFromSignature = Buffer.from(sha1, 'hex');
+
+    const hashBufferFromBody = crypto
+      .createHmac('sha1', this._appSecret)
+      .update(rawBody, 'utf8')
+      .digest();
+
+    // return early here if buffer lengths are not equal since timingSafeEqual
+    // will throw if buffer lengths are not equal
+    if (bufferFromSignature.length !== hashBufferFromBody.length) {
+      return false;
+    }
+
+    return crypto.timingSafeEqual(bufferFromSignature, hashBufferFromBody);
+  }
+
+  preprocess({
+    method,
+    headers,
+    query,
+    rawBody,
+  }: RequestContext<RequestBody, { 'x-hub-signature'?: string }>) {
+    if (method.toLowerCase() === 'get') {
+      if (
+        query['hub.mode'] === 'subscribe' &&
+        query['hub.verify_token'] === this.verifyToken
+      ) {
+        return {
+          shouldNext: false,
+          response: {
+            status: 200,
+            body: query['hub.challenge'],
+          },
+        };
+      }
+
+      return {
+        shouldNext: false,
+        response: {
+          status: 403,
+          body: 'Forbidden',
+        },
+      };
+    }
+
+    if (method.toLowerCase() !== 'post') {
+      return {
+        shouldNext: true,
+      };
+    }
+
+    if (
+      headers['x-hub-signature'] &&
+      this.verifySignature(rawBody, headers['x-hub-signature'])
+    ) {
+      return {
+        shouldNext: true,
+      };
+    }
+
+    const error = {
+      message: 'Facebook Signature Validation Failed!',
+      request: {
+        rawBody,
+        headers: {
+          'x-hub-signature': headers['x-hub-signature'],
+        },
+      },
+    };
+
+    return {
+      shouldNext: false,
+      response: {
+        status: 400,
+        body: { error },
+      },
+    };
+  }
+}
